iss: Moved cosim CLI to vadl-cosim.py
